make docker easier to use for everyday development
Strip out adding the repo to the image and rather
mount the code to a volume in the container.

Pass CMD in on container invocating rather than
as a startup command for the image so that if the
dev is having issues with container start up, they
can debug it with a different start up command.

separate docker build and run so that it is not
rebuilt every time a container is started