install dependencies on host via docker container
Before we were installing dependencies on image
but then overwriting them by mounting the repo
over it on the container.

Now we are building the image without node_modules.
Then on the initial container invocation, we
run `npm i` which installs deps on the mounted
volume (binding them to the host machine) so
that they may be persisted for future containers