DX-392: nightwatch tests (#270)
* DX-392: use proxy for landedcost 'try it now'

* DX-392: fix nightwatch tests for deep equal

Was using JSON.stringify before and comparing
strings , however, since stringify order is not
consistent, it was not reliable.

Here Im using lodash.isEqual and ignoring null
and undefined values. This yields a boolean
that we assert. This is consistent but vague. So
if we need more informative logging for our
tests, I wrote a custom function (helpers/assert.js)
and commented it out that we can switch to in the
future.

Also, example for POST /v3/units was incorrect
so I updated example data in config

* DX-392: fix remaining nightwatch tests

Add custom deepEqual as a supplimentary test
to be run with lodash to better specify which
fields contrasted on failure

Exclude helpers from nightwatch tests

* DX-392: fix remaining get-started nightwatch tests

* DX-392: add test-web to ci

exclude helpers from nightwatch tests
(wasnt working before)

lint all tests

in test-web, skip running pre-tests
and building jekyll. This is because
in ci it was done on the previous task
so if you need this in local testing,
youll need to do this manually

* DX-392: add dockerfile for ci selenium testing

Dockerfile.ci still isnt building properly

.travis.yml needs a few more commands to set the
drivers and selenium .jar from the image into
so that nightwatch leverages them

* DX-392: use container to run selenium

* DX-392: revert back Dockerfile changes

realized I didnt need to build my own docker image

add a script to be mounted when using selenium chrome
container

* DX-392: remove selenium stuff from travis, too