DX-349: api endpoint restyle (#272)
* [PORT-5641] Update host for Avalara for Communications REST API. (#230) (#1)

* removed html comparison code

* removed html-proofer gem dependency

* DX-349: pull model names from $refs in swagger docs

This is to just get it working. From here would
be ideal to extract most of the new HTML out of
endpoint.jsx and into a separate React component

Reason for passing $refs in this way was because
when dereferencing swagger docs, it gets rid of
$ref value and theres no way of retrieving it
for the anchor in the api-reference page.
Therefore, we pass the docs twice, once dereferenced,
and once not so that we can refer to these $ref
tags. I could not find anything in Swagger docs
that supported something like this out of the box.

* DX-349: move markup into separate component

and delete components no longer being used

* DX-349: fix links for models

Links to models from api-reference work now
(if merged with DX-347)

* DX-349: add markdown rendering

* DX-349: extract apiDocumentationParam in separate component

* table jsx fixes, model links now account for array req/res

* Array[x] syntax not part of model anchor tag

* DX-349: one URL instead of sandbox and prod