DX-420: Enums (#299)
* DX-420: add link to enum in methods attributes column

* DX-420: enums working for model attributes

Building dynamic html in .js file because there is
a fair amount of logic to determine where in the
swagger docs enums are defined and include comments
and this would be intensive for the jekyll compiler
which would slow down our builds even more

* DX-420: generate links to parent models in title

* DX-420: get enum paths working for queryString & pathParams

Using less liquid rendering. Some slight restructuring of
object when buildEnumFromMethod so that building the
enum is agnostic to where param came from

* DX-420: add tests for enum pages

Wrap models and enum pages in div with class that
better describes what page is accomplishing. Will
be useful for testing

* DX-420: account for models of type enum

Some models are of type string (enum) and that
is the only property. So they are missing the
`properties` fields. This change addresses those
models and builds html files for them

* DX-420: write tests for CountryList enum

Also switch order of Enum and link in models page

Fix slight bug introduced in previous commit
in build-enums as well as order or table in
apiDocumentationParam. Add #id to models.html
overall div for better testing